rag_agent_pipeline/requirements.txt
langgraph>=0.3.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-classic
langchain-text-splitters>=0.2.0
flashrank>=0.2.0
langchain-qdrant>=0.1.0
qdrant-client>=1.9.0
pymupdf>=1.24.0
pytesseract>=0.3.10
pillow>=10.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pytest>=8.0.0
